7.1. Refactoring reducer.
