+++ src/components/private-route/private-route.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {Route, Redirect} from 'react-router-dom';
 import {connect} from 'react-redux';
+import {getAuthorizationStatus} from "../../store/user/selector";
 
 const PrivateRoute = ({render, path, exact, authorizationStatus}) => {
   return (
@@ -23,7 +24,7 @@ PrivateRoute.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  authorizationStatus: state.authorizationStatus,
+  authorizationStatus: getAuthorizationStatus(state),
 });
 
 const ConnectedPrivateRoute = connect(mapStateToProps)(PrivateRoute);
