+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import {App} from "./components/app/app";
 import {applyMiddleware, createStore} from "redux";
-import {reducer} from "./store/reducer";
+import {combinedReducer} from "./store/reducer";
 import {Provider} from "react-redux";
 import {composeWithDevTools} from "redux-devtools-extension";
 import {createAPI} from "./services/api";
@@ -10,6 +10,7 @@ import thunk from "redux-thunk";
 import {checkAuth, fetchFilmList} from "./store/api-actions";
 import {redirect} from "./store/redirect";
 import {LoadingScreen} from "./components/loading-screen/loading-screen";
+import {getFilmsByCurrentGenre, increaseShownFilmQuantity} from "./store/action";
 
 const title = `The Grand Budapest Hotel`;
 const genre = `Drama`;
@@ -18,7 +19,7 @@ const date = 2014;
 const promo = {title, genre, date};
 
 const api = createAPI();
-const store = createStore(reducer, composeWithDevTools(applyMiddleware(thunk.withExtraArgument(api)), applyMiddleware(redirect)));
+const store = createStore(combinedReducer, composeWithDevTools(applyMiddleware(thunk.withExtraArgument(api)), applyMiddleware(redirect)));
 
 Promise.resolve()
   .then(() => {
@@ -26,6 +27,8 @@ Promise.resolve()
   })
   .then(() => store.dispatch(checkAuth()))
   .then(() => store.dispatch(fetchFilmList()))
+  .then(() => store.dispatch(getFilmsByCurrentGenre()))
+  .then(() => store.dispatch(increaseShownFilmQuantity()))
   .then(() => {
     ReactDOM.render(
         <Provider store={store}>
