+++ src/store/user/selector.js
@@ -1,0 +1,6 @@
+import {NameSpace} from "../reducer";
+
+const getAuthorizationStatus = (state) => state[NameSpace.USER].authorizationStatus;
+const getAuthInfo = (state) => state[NameSpace.USER].authInfo;
+
+export {getAuthorizationStatus, getAuthInfo};
